requirements.txt
absl-py==0.12.0
backcall==0.2.0
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
Cython==0.29.23
decorator==5.0.7
Flask==1.1.2
Flask-Cors==3.0.10
google-auth==1.30.0
google-auth-oauthlib==0.4.4
grpcio==1.37.1
idna==2.10
importlib-metadata==4.0.1
ipython==7.23.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.1
matplotlib-inline==0.1.2
numpy==1.20.2
oauthlib==3.1.0
opencv-python==4.5.1.48
paho-mqtt==1.5.1
pandas==1.2.4
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.3
seaborn==0.11.1
six==1.16.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
thop==0.0.31.post2005241907
torch==1.8.1
torchvision==0.9.1
tqdm==4.60.0
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.4.1
